hotel-standards-assistant/requirements.txt

streamlit
PyMuPDF
faiss-cpu
sentence-transformers
openai
googletrans==4.0.0rc1
numpy
